display images refactor complete
