+++ vite-project/src/Modules/DisplayImages/hooks/useFetchFilterImages.ts
@@ -1,14 +1,14 @@
-import { InfiniteData, useInfiniteQuery } from "@tanstack/react-query";
+import { InfiniteData } from "@tanstack/react-query";
 import { debounce } from "lodash";
 import { useEffect, useState } from "react";
 import { PostsResponse, QueryFetchMode, SinglePost } from "../lib/types";
-import { fetchPosts } from "../../../lib/api";
 import { useAuth0 } from "@auth0/auth0-react";
 import { useCustomInfiniteQuery } from "./useCustomInfiniteQuery";
 import { useSelector } from "react-redux";
 import { selectUser } from "../../../redux/selectors";
+import { useLocation } from "react-router-dom";
 
-export const useFetchFilterImages = (mode: QueryFetchMode) => {
+export const useFetchFilterImages = () => {
   const { isAuthenticated, user } = useAuth0();
   const { favourites } = useSelector(selectUser);
 
@@ -20,30 +20,24 @@ export const useFetchFilterImages = (mode: QueryFetchMode) => {
     null
   );
 
+  const { pathname } = useLocation();
+
+  const [mode, setMode] = useState<QueryFetchMode>(
+    pathname === "/my-posts"
+      ? QueryFetchMode.MY_POSTS
+      : pathname === "/"
+      ? QueryFetchMode.POSTS
+      : QueryFetchMode.FAVOURITES
+  );
+
+  useEffect(() => {
+    //useEffect require for when navigating from /my-posts to /favourites, as it uses the same component
+    if (pathname === "/favourites") setMode(QueryFetchMode.FAVOURITES);
+    if (pathname === "/my-posts") setMode(QueryFetchMode.MY_POSTS);
+  }, [pathname]);
+
   const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } =
     useCustomInfiniteQuery({ mode, user, favourites });
-
-  // const { data, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } =
-  //   useInfiniteQuery<PostsResponse>({
-  //     queryKey: ["posts"],
-  //     queryFn: ({ pageParam = 1 }) => fetchPosts({ pageParam, pageSize: 10 }),
-  //     getNextPageParam: (lastPage) => {
-  //       if (
-  //         lastPage.currentPage !== undefined &&
-  //         lastPage.totalPages !== undefined
-  //       ) {
-  //         if (lastPage.currentPage < lastPage.totalPages) {
-  //           return lastPage.currentPage + 1;
-  //         }
-  //       } else {
-  //         return undefined;
-  //       }
-  //     },
-  //     onError(err) {
-  //       console.log(err);
-  //       alert("Opps Something went wrong, Please try again later");
-  //     },
-  //   });
 
   const postData = data as InfiniteData<PostsResponse>;
   const allPostsz = postData?.pages.flatMap((page) => page.data) ?? [];
@@ -98,5 +92,6 @@ export const useFetchFilterImages = (mode: QueryFetchMode) => {
     allPostsz,
     isAuthenticated,
     user,
+    mode,
   };
 };
