+++ vite-project/src/Modules/DisplayImages/pages/Home.tsx
@@ -5,7 +5,6 @@ import Button from "../../../components/Button";
 import AnimatedWrapper from "../../../components/Containers/AnimatedWrapper";
 import { RenderCards } from "../components/Cards";
 import { useFetchFilterImages } from "../hooks/useFetchFilterImages";
-import { QueryFetchMode } from "../lib/types";
 
 const Home = () => {
   const {
@@ -16,7 +15,7 @@ const Home = () => {
     isFetchingNextPage,
     allPostsz,
     isAuthenticated,
-  } = useFetchFilterImages(QueryFetchMode.POSTS);
+  } = useFetchFilterImages();
 
   return (
     <AnimatedWrapper>
